Firebase storage, cloud messaging

+++ app/src/main/java/com/kotlin/android/inbyulgram/Feed.kt
@@ -1,10 +1,10 @@
 package com.kotlin.android.inbyulgram
 
-class Feed (
+class Feed(
     val userId:String,
     val imageUrl:String,
     val profileImageUrl:String,
-    var likeCount: Int,
+    var likeCount: Long,
     var isLike: Boolean,
     var isBookmark: Boolean
 )
